add js for phone numbers

+++ db/migrate/20210507160857_add_name_email_and_phone_number_to_spree_producers.rb
@@ -2,6 +2,6 @@ class AddNameEmailAndPhoneNumberToSpreeProducers < ActiveRecord::Migration[6.0]
   def change
     add_column :spree_producers, :contact_name, :string
     add_column :spree_producers, :email, :string
-    add_column :spree_producers, :phone, :integer
+    add_column :spree_producers, :phone, :bigint
   end
 end
